Fix placing of `pretest-*` CSS classes
If a contest is not marked as `run_pretests_only`, prestests will not be run, even if `is_pretested` is true for a contest problem. However, not all contest setters diable both flags once a problem has been system tested, which gives the false impression that the problem was not system tested yet.